heuristic works but far from perfect

+++ src/shared/state/Deck.cpp
@@ -95,5 +95,5 @@ std::vector<int> Deck::CountCards(){
 }
 
 bool Deck::IsFull(){
-  return (cards.size() == sizeMax);
+  return ((int) cards.size() == sizeMax);
 }
